Transport: accept_ack after init.
In the ranch_ssl implementation, calling accept_ack will try to trigger
a ssl_handshake via gen_server calls to the ssl socket. However since
the socket is still stuck in init() it results in a deadlock.

To fix this, start the gen_server with a timeout of 0. That way the
timeout will trigger immediately after init(), and we can run init_ack()
there.